requirements.txt
Django==1.8.13
ipython==5.1.0
python-memcached==1.58
psycopg2==2.6.1
redis==2.10.5
django-celery==3.1.17
gunicorn==19.6.0
django-memcache-admin==0.2.1
django-haystack==2.5.0
elasticsearch==2.4.0
